+++ packages/cli/eslint.config.mjs
@@ -1,0 +1,34 @@
+import baseConfig from "../../eslint.config.mjs";
+
+export default [
+    ...baseConfig,
+    {
+        "files": [
+            "**/*.json"
+        ],
+        "rules": {
+            "@nx/dependency-checks": [
+                "error",
+                {
+                    "ignoredFiles": [
+                        "{projectRoot}/eslint.config.{js,cjs,mjs,ts,cts,mts}"
+                    ],
+                    "ignoredDependencies": [
+                        "vitest",
+                        "viem",
+                        "tsup",
+                        "tslib"
+                    ]
+                }
+            ]
+        },
+        "languageOptions": {
+            "parser": (await import('jsonc-eslint-parser'))
+        }
+    },
+    {
+        ignores: [
+            "**/out-tsc"
+        ]
+    }
+];