+++ packages/ui/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.2 (2026-03-10)
+
+### 🧱 Updated Dependencies
+
+- Updated @jaw.id/core to 0.1.2
+
 ## 0.1.1 (2026-03-04)
 
 ### 🧱 Updated Dependencies